Refactored files for supporting planned api system additions
