--- app/api/www/index.js
@@ -1,8 +1,0 @@
-export function fetchUrl({ url }) {
-  return fetch(`${url}`).then((res) => {
-    if (res.status === 200) {
-      return res.json();
-    }
-    return null;
-  });
-}
