+++ app/utils/log.js
@@ -2,7 +2,7 @@ import os, { EOL } from 'os';
 import { IS_DEV } from '../constants/env';
 import { APP_NAME, APP_VERSION } from '../constants/meta';
 import { PATHS } from './paths';
-import { appendFileAsync } from '../api/sys/fileOps';
+import { appendFileAsync } from '../data/sys/fileOps';
 import { dateTimeUnixTimestampNow } from './date';
 
 const { logFile } = PATHS;
